876
Wow, this one really got me. My solution was so convoluted, but the optimal solution approach two is so clean.

Also, approach two is preferred over approach one, because approach one uses an array, which defeats the purpose of the linked list. Whereas approach two uses the linked list and demonstrates an understanding of the structure and how to use it.