end of 27: remove element
The editorial solutions don't actually remove the elements. In the example [3,2,2,3] the last 3 is still left in the array. Both solutions do this. What is the point? Could easily just loop and count each element not equal to val. My solution at least bubbles the elements up and replaces them with zeros.